create semester block selector for mobile and desktop
